convert test-utils to typescript (#7877)
* move js utils into directory


* typescripterize js test-utils


* move test utils to directory


* move remaining mock functions to directory


* update imports


* missed copyright


* type wait for update
