+++ test/components/views/spaces/SpaceSettingsVisibilityTab-test.tsx
@@ -11,7 +11,7 @@ import { GuestAccess, HistoryVisibility, JoinRule } from 'matrix-js-sdk/src/@typ
 
 import _SpaceSettingsVisibilityTab from "../../../../src/components/views/spaces/SpaceSettingsVisibilityTab";
 import { createTestClient, mkEvent, wrapInMatrixClientContext } from '../../../test-utils';
-import { mkSpace, mockStateEventImplementation } from '../../../utils/test-utils';
+import { mkSpace, mockStateEventImplementation } from '../../../test-utils';
 import { MatrixClientPeg } from '../../../../src/MatrixClientPeg';
 
 const SpaceSettingsVisibilityTab = wrapInMatrixClientContext(_SpaceSettingsVisibilityTab);
